Move unit tests into separate module
