+++ src/elementary.rs
@@ -76,46 +76,3 @@ pub fn lnf(x: f32) -> f32 {
         panic!("computing lnf({x}) needs at least one of the `std` or `libm` feature flags to be enabled");
     }
 }
-
-#[cfg(test)]
-mod test {
-    use approx::{assert_abs_diff_eq, assert_relative_eq};
-
-    use super::*;
-
-    #[test]
-    fn sanity_check_log() {
-        assert!(ln(-1.0).is_nan());
-        assert_abs_diff_eq!(ln(1.0), 0.0);
-        assert_abs_diff_eq!(ln(2.0), core::f64::consts::LN_2);
-        assert_abs_diff_eq!(ln(f64::MAX), 709.782_712_893_384);
-    }
-
-    #[test]
-    fn sanity_check_logf() {
-        assert!(lnf(-1.0).is_nan());
-        assert_abs_diff_eq!(lnf(1.0), 0.0);
-        assert_abs_diff_eq!(lnf(2.0), core::f32::consts::LN_2);
-        assert_abs_diff_eq!(lnf(f32::MAX), 88.722_839);
-    }
-
-    #[test]
-    fn sanity_check_sqrt() {
-        assert!(sqrt(-1.0).is_nan());
-        assert_abs_diff_eq!(sqrt(0.0), 0.0);
-        assert_abs_diff_eq!(sqrt(1.0), 1.0);
-        assert_abs_diff_eq!(sqrt(2.0), 1.414_213_562_373_095_1);
-        assert_abs_diff_eq!(sqrt(4.0), 2.0);
-        assert_abs_diff_eq!(sqrt(f64::MAX), 1.340_780_792_994_259_6e154);
-    }
-
-    #[test]
-    fn sanity_check_sqrtf() {
-        assert!(sqrtf(-1.0).is_nan());
-        assert_abs_diff_eq!(sqrtf(0.0), 0.0);
-        assert_abs_diff_eq!(sqrtf(1.0), 1.0);
-        assert_abs_diff_eq!(sqrtf(2.0), 1.414_213_5);
-        assert_abs_diff_eq!(sqrtf(4.0), 2.0);
-        assert_relative_eq!(sqrtf(f32::MAX), 1.844_674_4e19);
-    }
-}
