+++ src/lib.rs
@@ -126,6 +126,8 @@ mod sw0;
 mod sw0f;
 mod swm1;
 mod swm1f;
+#[cfg(test)]
+mod unit_tests;
 
 /// The negative inverse of e (-1/e).
 ///
